Massive update. New levels/content and dual waves

+++ src/com/ggstudios/divisionbyzero/Clickable.java
@@ -1,7 +1,5 @@
 package com.ggstudios.divisionbyzero;
 
-import android.view.MotionEvent;
-
 public interface Clickable {
-	public boolean onTouchEvent(int action, int x, int y);
+	public boolean onTouchEvent(int action, float x, float y);
 }
